access controll

+++ app/Http/Controllers/PostsController.php
@@ -12,7 +12,23 @@ class PostsController extends Controller
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
+     * 
+     * 
      */
+
+
+         /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth',['except' => ['index', 'show']]);
+    }
+
+
+
     public function index()
     {
         // $posts = Post::all();
@@ -78,6 +94,11 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error','Unauthorized access');
+        }
+
         return view('posts.edit')->with('post',$post);
     }
 
@@ -112,6 +133,13 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error','Unauthorized access');
+        }
+
+
+
         $post->delete();
         return redirect('/posts')->with('success', 'Post Deleted');
     }
